+++ src/senaite/core/upgrade/v02_06_000.py
@@ -1712,6 +1712,17 @@ def migrate_samplepoints_coordinates(tool):
     logger.info("Migrating coordinates from SamplePoint [DONE]")
 
 
+def remove_category_title_metadata(tool):
+    """Remove getCategoryTitle metadata from catalogs
+    """
+    logger.info("Removing getCategoryTitle metadata from catalogs ...")
+
+    del_metadata(ANALYSIS_CATALOG, "getCategoryTitle")
+    del_metadata(SETUP_CATALOG, "getCategoryTitle")
+
+    logger.info("Removing getCategoryTitle metadata from catalogs [DONE]")
+
+
 def set_referenceable_behavior(tool):
     """Assigns the referenceable behavior to setup folders so they are indexed
     in the UID Catalog